+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 =================
 
 ## 0.18.50
+### is.codion.common.core
+- ValueCollection.optional() now returns an empty Optional if the collection is empty.
 ### is.codion.common.model
 - DefaultConditionModel.DefaultSetCondition.equals() and hashCode() now throw UnsupportedOperationException, in case equalTo() is confused with equals().
 - DefaultConditionModel bug fixed, adding wildards no longer breaks case-insensitivity.